build: icon-component
to reduce the amount of imports the icon component will be used to import

#233
